Added shortcut
this shortcut makes it possible to get the Spotify URL for any song, album or record you find on Apple Music. this makes it easier to share music with people who use Spotify.

OBS: this shortcut relies on the Odesli API to work. if that goes down at some point in the future (which i doubt is soon because it’s stable and used widely), then the Shortcut will cease to work. i might update if that’s the case at some point in the future.